chore(compose): split docker compose into 2 pieces, with separate dotfiles
For organization purposes, and in the spirit of transitioning into kubernetes style yaml organization, the docker-compse files were split into 2, one solely for provisioning the database, and the other solely for the rocket service. The dotfiles were also split for better organization. The readme was updated to reflect those changes.